# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/quality_inspection_readings:
    get:
      tags:
        - QualityInspectionReadings
      summary: List QualityInspectionReadings
      description: Get a paginated list of QualityInspectionReadings with optional filtering and sorting
      operationId: listQualityInspectionReadings
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QualityInspectionReadingList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - QualityInspectionReadings
      summary: Create QualityInspectionReading
      description: Create a new QualityInspectionReading
      operationId: createQualityInspectionReading
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQualityInspectionReadingRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QualityInspectionReading'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/quality_inspection_templates:
    get:
      tags:
        - QualityInspectionTemplates
      summary: List QualityInspectionTemplates
      description: Get a paginated list of QualityInspectionTemplates with optional filtering and sorting
      operationId: listQualityInspectionTemplates
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QualityInspectionTemplateList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - QualityInspectionTemplates
      summary: Create QualityInspectionTemplate
      description: Create a new QualityInspectionTemplate
      operationId: createQualityInspectionTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQualityInspectionTemplateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QualityInspectionTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    QualityInspectionReading:
      type: object
      required:
        - id
        - company_id
        - inspection_id
        - parameter_name
        - numeric
        - result
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        inspection_id:
          type: string
          format: uuid
        parameter_name:
          type: string
          maxLength: 140
        numeric:
          type: boolean
        reading_value:
          nullable: true
          type: string
        min_value:
          nullable: true
          type: string
        max_value:
          nullable: true
          type: string
        manual_result:
          nullable: true
          type: boolean
        result:
          type: string
        remarks:
          nullable: true
          type: string
          maxLength: 280
        metadata:
          type: object
...
    CreateQualityInspectionReadingRequest:
      type: object
      required:
        - company_id
        - inspection_id
        - parameter_name
        - numeric
        - result
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        inspection_id:
          type: string
          format: uuid
        parameter_name:
          type: string
          maxLength: 140
        numeric:
          type: boolean
        reading_value:
          nullable: true
          type: string
        min_value:
          nullable: true
          type: string
        max_value:
          nullable: true
          type: string
        manual_result:
          nullable: true
          type: boolean
        result:
          type: string
        remarks:
          nullable: true
          type: string
          maxLength: 280
        metadata:
          type: object
...
    QualityInspectionReadingList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/QualityInspectionReading'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    QualityInspectionTemplate:
      type: object
      required:
        - id
        - company_id
        - template_name
        - is_active
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        template_name:
          type: string
          maxLength: 140
        item_id:
          nullable: true
          type: string
          format: uuid
        is_active:
          type: boolean
        metadata:
          type: object
...
    CreateQualityInspectionTemplateRequest:
      type: object
      required:
        - company_id
        - template_name
        - is_active
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        template_name:
          type: string
          maxLength: 140
        item_id:
          nullable: true
          type: string
          format: uuid
        is_active:
          type: boolean
        metadata:
          type: object
...
    QualityInspectionTemplateList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/QualityInspectionTemplate'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []