# src/presentation/http/quality_inspection_template_handler.rs
//! QualityInspectionTemplate REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{QualityInspectionTemplateService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateQualityInspectionTemplateDto, UpdateQualityInspectionTemplateDto, PatchQualityInspectionTemplateDto, QualityInspectionTemplateResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum QualityInspectionTemplateError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for QualityInspectionTemplateError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for QualityInspectionTemplateError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "QUALITYINSPECTIONTEMPLATE_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "QUALITYINSPECTIONTEMPLATE_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "QUALITYINSPECTIONTEMPLATE_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "QUALITYINSPECTIONTEMPLATE_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for QualityInspectionTemplate.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /quality_inspection_templates | List with pagination |
/// | POST | /quality_inspection_templates | Create new |
/// | GET | /quality_inspection_templates/:id | Get by ID |
/// | PUT | /quality_inspection_templates/:id | Full update |
/// | PATCH | /quality_inspection_templates/:id | Partial update |
/// | DELETE | /quality_inspection_templates/:id | Soft delete |
/// | POST | /quality_inspection_templates/bulk | Bulk create |
/// | POST | /quality_inspection_templates/upsert | Upsert |
/// | GET | /quality_inspection_templates/trash | List deleted |
/// | POST | /quality_inspection_templates/:id/restore | Restore |
/// | DELETE | /quality_inspection_templates/empty | Empty trash |
/// | GET | /quality_inspection_templates/:id/deleted | Get deleted by ID |
/// | DELETE | /quality_inspection_templates/trash/:id | Permanent delete from trash |
/// | GET | /quality_inspection_templates/count | Count active entities |
/// | GET | /quality_inspection_templates/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(QualityInspectionTemplateService::with_repository(repository));
/// let router = create_quality_inspection_template_routes(service);
/// ```
pub fn create_quality_inspection_template_routes(service: Arc<QualityInspectionTemplateService>) -> Router {
    BackboneCrudHandler::<QualityInspectionTemplateService, QualityInspectionTemplate, CreateQualityInspectionTemplateDto, UpdateQualityInspectionTemplateDto, QualityInspectionTemplateResponseDto>::routes(
        service,
        "/quality_inspection_templates",
    )
}

/// Create Axum router with only the read (GET) endpoints for QualityInspectionTemplate.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_quality_inspection_template_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_quality_inspection_template_read_routes(service: Arc<QualityInspectionTemplateService>) -> Router {
    BackboneCrudHandler::<QualityInspectionTemplateService, QualityInspectionTemplate, CreateQualityInspectionTemplateDto, UpdateQualityInspectionTemplateDto, QualityInspectionTemplateResponseDto>::read_routes(
        service,
        "/quality_inspection_templates",
    )
}

/// Create Axum router with only the write (mutation) endpoints for QualityInspectionTemplate.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
pub fn create_quality_inspection_template_write_routes(service: Arc<QualityInspectionTemplateService>) -> Router {
    BackboneCrudHandler::<QualityInspectionTemplateService, QualityInspectionTemplate, CreateQualityInspectionTemplateDto, UpdateQualityInspectionTemplateDto, QualityInspectionTemplateResponseDto>::write_routes(
        service,
        "/quality_inspection_templates",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_quality_inspection_template_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<QualityInspectionTemplateService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_quality_inspection_template_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}